requirements.txt
fastapi
uvicorn
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]==1.7.4